Create Function to Determine a Winner
Declared pickWinner(), which takes two functions as paramaters.
The left paramater takes in the user's choice and the right
parameter takes in the computer's choice.

pickWinner() uses the user's choice to compare against the computers
choice with if else statements, to then determine a winner. The function
ends with a call to the original function, essentially restarting the
game.